Support multiple config schemas, use Cobra
