+++ go.mod
@@ -1,5 +1,15 @@
-module github.com/snarlysodboxer/try-yaml-v3
+module github.com/snarlysodboxer/predictable-yaml
 
 go 1.17
 
-require gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+require (
+	github.com/spf13/cobra v1.5.0
+	gopkg.in/yaml.v3 v3.0.0
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/kr/pretty v0.3.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+)
