+++ test-data/service.invalid.yaml
@@ -1,0 +1,17 @@
+---
+apiVersion: v1
+kind: Service
+metadata:
+  namespace: example
+  name: example
+  labels:
+    asdf: example
+spec:
+  type: ClusterIP
+  ports:
+  - protocol: TCP
+    name: example
+    port: 8080
+    targetPort: example
+  selector:
+    app: example
