+++ test-data/service.valid.yaml
@@ -1,0 +1,17 @@
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: example
+  namespace: example
+  labels:
+    app: example
+spec:
+  type: ClusterIP
+  selector:
+    app: example
+  ports:
+  - name: example
+    protocol: TCP
+    port: 8080
+    targetPort: example
